MRD-2804 Move clean-up task to new package
